@@ -1,24 +1,20 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
   </PropertyGroup>
 
   <ItemGroup>
-    <None Remove="App_Data\Config\service.grpc.json~RF2d8f7391.TMP" />
-    <None Remove="App_Data\Config\service.netty.json~RF2dbfd966.TMP" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Akka.Remote" Version="1.3.16" />
+    <PackageReference Include="Akka.Remote" Version="1.3.17" />
     <PackageReference Include="Microsoft.AspNetCore.Mvc" Version="2.2.0" />
     <PackageReference Include="Microsoft.AspNetCore.Mvc.Core" Version="2.2.5" />
-    <PackageReference Include="Microsoft.Extensions.Hosting" Version="3.0.1" />
+    <PackageReference Include="Microsoft.Extensions.Hosting" Version="3.1.0" />
     <PackageReference Include="NLog" Version="4.6.8" />
     <PackageReference Include="NLog.Extensions.Hosting" Version="1.6.1" />
     <PackageReference Include="NLog.Extensions.Logging" Version="1.6.1" />
-    <PackageReference Include="SoapCore" Version="0.9.9.6" />
+    <PackageReference Include="SoapCore" Version="1.0.0" />
+    <PackageReference Include="System.Text.Encoding.CodePages" Version="4.7.0" />
   </ItemGroup>
 
   <ItemGroup>
